+++ spec/logstash/plugin_mixins/ecs_compatibility_support_spec.rb
@@ -46,8 +46,11 @@ describe LogStash::PluginMixins::ECSCompatibilitySupport do
             plugin_class.send(:include, ecs_compatibility_support)
           end
 
-          it 'supports an `ecs_compatibility` option' do
+          it 'supports an `ecs_compatibility` config option' do
             expect(plugin_class.get_config).to include('ecs_compatibility')
+          end
+
+          it 'defines an `ecs_compatibility` method' do
             expect(plugin_class.method_defined?(:ecs_compatibility)).to be true
           end
 
@@ -63,12 +66,15 @@ describe LogStash::PluginMixins::ECSCompatibilitySupport do
             end
 
             # TODO: Remove once ECS Compatibility config is included in one or
-            # more Logstash release branches. This speculative spec is meant to
-            # run on Logstashes prior to the introduction of a core implementation.
+            #       more Logstash release branches. This speculative spec is meant
+            #       to prove that this implementation will not override an existing
+            #       implementation.
             context 'if base class were to include ecs_compatibility config' do
               let(:plugin_base_class) do
                 Class.new(super()) do
-                  config :ecs_compatibility, :validate => :boolean, :default => false
+                  config :ecs_compatibility
+                  def ecs_compatibility
+                  end
                 end
               end
               before(:each) do
@@ -80,34 +86,53 @@ describe LogStash::PluginMixins::ECSCompatibilitySupport do
 
           # The four plugin base classes override their own `#initialize` to also
           # send `#config_init`, so we can count on the options being normalized
-          # and populated out to the relevant ivars.
+          # and available.
           context 'when initialized' do
             let(:plugin_options) { Hash.new }
             subject(:instance) { plugin_class.new(plugin_options) }
 
-            context 'with `ecs_compatibility => true`' do
-              let(:plugin_options) { super().merge('ecs_compatibility' => 'true') }
-              its(:ecs_compatibility) { should be true }
-              it 'populates the @ecs_compatibility ivar with `true`' do
-                expect(instance.send(:instance_variable_get, :@ecs_compatibility)).to be true
-              end
+            context 'with `ecs_compatibility => v1`' do
+              let(:plugin_options) { super().merge('ecs_compatibility' => 'v1') }
+              its(:ecs_compatibility) { should equal :v1 }
             end
 
-            context 'with `ecs_compatibility => false`' do
-              let(:plugin_options) { super().merge('ecs_compatibility' => 'false') }
-              its(:ecs_compatibility) { should be false }
-              it 'populates the @ecs_compatibility ivar with `false`' do
-                expect(instance.send(:instance_variable_get, :@ecs_compatibility)).to be false
+            context 'with `ecs_compatibility => disabled`' do
+              let(:plugin_options) { super().merge('ecs_compatibility' => 'disabled') }
+              its(:ecs_compatibility) { should equal :disabled }
+            end
+
+            context 'with an invalid value for `ecs_compatibility`' do
+              shared_examples 'invalid value' do |invalid_value|
+                before { allow(plugin_class).to receive(:logger).and_return(logger_stub) }
+                let(:logger_stub) { double('Logger').as_null_object }
+
+                let(:plugin_options) { super().merge('ecs_compatibility' => invalid_value) }
+
+                it 'fails to initialize and emits a helpful log message' do
+                  # we cannot rely on internal details of the error that is emitted such as its exact message,
+                  # but we can expect the given value to be included in a message logged at ERROR-level.
+                  expect { plugin_class.new(plugin_options) }.to raise_error(LogStash::ConfigurationError)
+                  expect(logger_stub).to have_received(:error).with(/\b#{Regexp.escape(invalid_value.to_s)}\b/)
+                end
+              end
+
+              context('a random string') do
+                include_examples 'invalid value', 'bananas'
+              end
+
+              context('nil') do
+                include_examples 'invalid value', nil
+              end
+
+              context('an integer') do
+                include_examples 'invalid value', 17
               end
             end
 
             # we only specify default behaviour in cases where native support is _NOT_ provided.
             unless native_support_for_ecs_compatibility
               context 'without an `ecs_compatibility` directive' do
-                its(:ecs_compatibility) { should be false }
-                it 'populates the @ecs_compatibility ivar with `false`' do
-                  expect(instance.send(:instance_variable_get, :@ecs_compatibility)).to be false
-                end
+                its(:ecs_compatibility) { should equal :disabled }
               end
             end
           end