+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.0.0
 
- - Implements boolean `ecs_compatibility` config option, unless already provided by Logstash core.
+ - Support Mixin for ensuring a plugin has an `ecs_compatibility` method that is configurable from an `ecs_compatibility` option that accepts the literal `disabled` or an integer representing a major ECS version, using the implementation from Logstash core if available.